create metrics for different types of behavior
